+++ ui/web/src/adapters/chat-message.adapter.ts
