+++ ui/web/src/pages/traces/trace-detail-dialog.tsx
@@ -8,7 +8,8 @@ import { useWsEvent } from "@/hooks/use-ws-event";
 import { Events } from "@/api/protocol";
 import { formatDate, formatDuration, formatTokens, computeDurationMs } from "@/lib/format";
 import { useUiStore } from "@/stores/use-ui-store";
-import { SpanTreeNode, StatusBadge, buildSpanTree } from "./trace-span-tree-node";
+import { SpanTreeNode, StatusBadge } from "./trace-span-tree-node";
+import { buildSpanTree } from "@/adapters/trace.adapter";
 import { TracePreviewBlock } from "./trace-preview-block";
 import type { TraceData, SpanData } from "./hooks/use-traces";
 import type { AgentEventPayload } from "@/types/chat";
